+++ src/app/shared/components/loading-overlay/loading-overlay.component.ts
@@ -1,4 +1,4 @@
-import { Component, input } from '@angular/core';
+import { Component, input, ChangeDetectionStrategy } from '@angular/core';
 
 import { MatIconModule } from '@angular/material/icon';
 import { TranslateModule } from '@ngx-translate/core';
@@ -9,6 +9,7 @@ import { TranslateModule } from '@ngx-translate/core';
   imports: [MatIconModule, TranslateModule],
   templateUrl: './loading-overlay.component.html',
   styleUrls: ['./loading-overlay.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class LoadingOverlayComponent {
   title = input('shared.loadingOverlay.defaultTitle');